+++ tests/test_pdfalyze.py
@@ -51,12 +51,17 @@ def test_pdfalyze_CLI_streams_scan(adobe_type1_fonts_pdf_path):
     _assert_args_yield_lines(135, adobe_type1_fonts_pdf_path, '-s', '48')
 
 
+def test_pdfalyze_non_zero_return_code(SF424_page2_pdf_path):
+    with pytest.raises(CalledProcessError):
+        check_output([PDFALYZE, SF424_page2_pdf_path, '-t'], env=environ).decode()
+
+
 def test_yara_rules_option(adobe_type1_fonts_pdf_path, additional_yara_rules_path):
     _assert_args_yield_lines(2593, adobe_type1_fonts_pdf_path, '-Y', additional_yara_rules_path)
     _assert_args_yield_lines(1797, adobe_type1_fonts_pdf_path, '--no-default-yara-rules', '-Y', additional_yara_rules_path)  # noqa: E501
 
 
-@pytest.mark.slow
+# @pytest.mark.slow
 def test_quote_extraction(adobe_type1_fonts_pdf_path):
     _assert_args_yield_lines(2914, adobe_type1_fonts_pdf_path, '--extract-quoted', 'backtick', '-s')
     _assert_args_yield_lines(5736, adobe_type1_fonts_pdf_path, '--extract-quoted', 'backtick', '--extract-quoted', 'frontslash', '-s')  # noqa: E501
@@ -75,6 +80,7 @@ def _assert_args_yield_lines(line_count, file, *args):
 
 def _run_with_args(pdf, *args) -> str:
     """check_output() technically returns bytes so we decode before returning STDOUT output"""
+    args = list(args) + ['--allow-missed-nodes']
     return check_output([PDFALYZE, pdf, *args], env=environ).decode()